pyproject.toml
[tool.poetry]
name = "ithome-iron-2020-django"
version = "0.1.0"
description = ""
authors = ["Yan-ren Tsai <elleryq@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.1"
django-environ = "^0.4.5"
djangorestframework = "^3.11.1"
asgiref = "^3.2.10"
django-extensions = "^3.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
